Add basic Event scripting capaibilities
This implements the first event script, going from the first Ul'dah opening
cutscene to the next.

+++ src/world/ipc/mod.rs
@@ -337,7 +337,8 @@ pub enum ClientZoneIpcData {
     #[br(pre_assert(*magic == ClientZoneIpcType::EventRelatedUnk))]
     EventRelatedUnk {
         unk1: u32,
-        unk2: u32,
+        unk2: u16,
+        #[brw(pad_before = 2)]
         unk3: u32,
         unk4: u32,
     },
